Rebase the whole stack from any of its layers
- The bottom of a stack has no parent, so it counted as a lone
  branch: Rebase rewrote it on its own and only then restacked, and
  the stack derived after the rewrite no longer reached the branches
  above, whose merge base with the new commits sits on the default
  branch. Signing the bottom of an unopened stack lost the rest.
- `rebaseSigned` now hands a stacked worktree to the stack's restack
  from any layer, which records every tip before anything moves, and
  the footer shows the stack's Rebase on the bottom layer too. The
  restack pushes the published branches back up, as the lone path
  did and the documentation already said, and says Signed when that
  is all it did.
